Convert coordinates from `Float`s to `Int`s.

+++ src/SimulatedAnnealing/City.hs
@@ -7,13 +7,13 @@ where
 
 import           System.Random
 
-type City = (Float, Float)
+type City = (Int, Int)
 
-mkCity :: Float -> Float -> City
+mkCity :: Int -> Int -> City
 mkCity x y = (x, y)
 
 generateRandomCity :: IO City
 generateRandomCity = do
-  x <- randomRIO (1 :: Float, 100)
-  y <- randomRIO (1 :: Float, 100)
+  x <- randomRIO (1 :: Int, 100)
+  y <- randomRIO (1 :: Int, 100)
   return (x, y)
